test(chat): split chat dictation spec across CI shards
